- Version 2

+++ app/src/main/java/org/kaizoku/otropelisplusmas/model/Change.java
@@ -1,7 +1,7 @@
 package org.kaizoku.otropelisplusmas.model;
 
 public class Change {
-    public int type;
+    public byte type;
     public String text;
     public Change() {
     }
